tests/integration/test_llm_integration_vcr.py
"""LLM integration tests with VCR for reproducibility."""

import pytest
import vcr
import chess
from pathlib import Path

from llm_chess_arena.player.llm import LLMPlayer, LLMConnector, GameArenaLLMMoveHandler


# Configure VCR
CASSETTE_DIR = Path(__file__).parent / "cassettes"
CASSETTE_DIR.mkdir(exist_ok=True)

# VCR configuration for reproducible tests
vcr_config = vcr.VCR(
    cassette_library_dir=str(CASSETTE_DIR),
    record_mode="once",  # Record if cassette doesn't exist, replay otherwise
    match_on=["method", "scheme", "host", "port", "path", "query"],
    filter_headers=[
        "authorization",
        "x-api-key",
        "api-key",
    ],  # Remove sensitive headers
    filter_post_data_parameters=["api_key"],
    decode_compressed_response=True,
    before_record_response=lambda response: {
        **response,
        "headers": {
            k: v
            for k, v in response["headers"].items()
            if k.lower() not in ["date", "x-request-id", "cf-ray"]
        },
    },
)


class TestLLMIntegrationVCR:
    """Reproducible LLM integration tests using VCR."""

    @vcr_config.use_cassette("openai_single_move.yaml")
    def test_openai_single_move__with_vcr__then_reproducible(self):
        """Test OpenAI move generation with recorded response."""
        connector = LLMConnector(
            model="gpt-4o-mini",
            temperature=0.0,  # Deterministic
            max_tokens=500,
        )
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(
            connector=connector,
            handler=handler,
            color="white",
            name="OpenAI-VCR",
            max_move_retries=3,
            num_votes=1,
        )

        board = chess.Board()
        decision = player(board)

        # This will be reproducible across runs
        assert decision.action == "move"
        assert decision.attempted_move in ["e2e4", "d2d4", "g1f3", "b1c3"]

    @vcr_config.use_cassette("llm_retry_illegal_move.yaml")
    def test_llm_retry_on_illegal__with_vcr__then_recovers(self):
        """Test LLM retry logic with recorded responses."""
        # Create a mock connector that returns illegal move first
        connector = LLMConnector(model="gpt-4o-mini", temperature=0.0)
        handler = GameArenaLLMMoveHandler()

        player = LLMPlayer(
            connector=connector,
            handler=handler,
            color="white",
            max_move_retries=5,  # Increased retries to handle stubborn models
            num_votes=1,
        )

        # Use standard starting position - simpler for LLM to handle
        board = chess.Board()

        decision = player(board)

        # Should eventually return a legal move
        assert decision.action == "move"
        # Verify the move is legal
        move = chess.Move.from_uci(decision.attempted_move)
        assert move in board.legal_moves

    @vcr_config.use_cassette("llm_majority_voting.yaml")
    def test_llm_majority_voting__with_vcr__then_deterministic(self):
        """Test majority voting with recorded responses."""
        connector = LLMConnector(
            model="gpt-4o-mini",
            temperature=0.7,  # Higher temp for variety
        )
        handler = GameArenaLLMMoveHandler()

        player = LLMPlayer(
            connector=connector,
            handler=handler,
            color="white",
            num_votes=3,  # Request 3 samples
        )

        board = chess.Board()
        decision = player(board)

        # Result should be deterministic with cassette
        assert decision.action == "move"
        assert decision.attempted_move is not None

    @vcr_config.use_cassette("llm_complex_position.yaml")
    def test_llm_complex_position__with_vcr__then_finds_good_move(self):
        """Test LLM on complex middlegame position with recording."""
        connector = LLMConnector(model="gpt-4o-mini", temperature=0.0)
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(connector=connector, handler=handler, color="white")

        # Complex tactical position
        board = chess.Board(
            "r1bqk2r/pp2bppp/2n1pn2/3p4/2PP4/2N1PN2/PP2BPPP/R1BQKR2 w Qkq - 2 9"
        )

        decision = player(board)

        assert decision.action == "move"
        # Move should be legal
        move = chess.Move.from_uci(decision.attempted_move)
        assert move in board.legal_moves

    @vcr_config.use_cassette("llm_endgame_position.yaml")
    def test_llm_endgame__with_vcr__then_handles_correctly(self):
        """Test LLM in endgame position with recording."""
        connector = LLMConnector(model="gpt-4o-mini", temperature=0.0)
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(connector=connector, handler=handler, color="white")

        # King and pawn endgame
        board = chess.Board("8/8/8/3k4/8/3K4/3P4/8 w - - 0 1")

        decision = player(board)

        assert decision.action == "move"
        assert decision.attempted_move in ["d3e3", "d3c3", "d2d4", "d3e2"]


class TestLLMErrorHandlingVCR:
    """Test error scenarios with VCR."""

    @vcr_config.use_cassette("llm_timeout_simulation.yaml")
    def test_llm_timeout__with_vcr__then_handles_gracefully(self):
        """Test timeout handling with recorded response."""
        connector = LLMConnector(
            model="gpt-4o-mini",
            timeout=1,  # Very short timeout
        )
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(connector=connector, handler=handler, color="white")

        board = chess.Board()

        # This test depends on whether timeout was recorded
        # In practice, you'd want to mock the timeout separately
        try:
            decision = player(board)
            # If it succeeds (from cassette), verify it's valid
            assert decision.action in ["move", "resign"]
        except Exception as e:
            # If it fails, verify it's a timeout-related error
            assert "timeout" in str(e).lower() or "time" in str(e).lower()

    @vcr_config.use_cassette("llm_malformed_response.yaml")
    def test_llm_malformed_response__with_vcr__then_retries(self):
        """Test handling of malformed LLM responses."""
        # This would need a cassette with actual malformed responses
        # For now, we test the retry mechanism
        connector = LLMConnector(model="gpt-4o-mini", temperature=1.0)
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(
            connector=connector, handler=handler, color="white", max_move_retries=3
        )

        board = chess.Board()
        decision = player(board)

        # Should eventually get a valid move despite potential malformed responses
        assert decision.action == "move"
        assert decision.attempted_move is not None


@pytest.fixture
def vcr_cassette_name(request):
    """Generate cassette name from test name."""
    return f"{request.node.name}.yaml"


@pytest.fixture
def llm_player_vcr(vcr_cassette_name):
    """Create LLM player with VCR recording."""
    with vcr_config.use_cassette(vcr_cassette_name):
        connector = LLMConnector(model="gpt-4o-mini", temperature=0.0)
        handler = GameArenaLLMMoveHandler()
        player = LLMPlayer(connector=connector, handler=handler, color="white")
        yield player


class TestLLMGamesVCR:
    """Test full games with VCR recording."""

    def test_llm_vs_random_game__with_vcr__then_completes(self, llm_player_vcr):
        """Test a full game between LLM and random player."""
        from llm_chess_arena.player.random_player import RandomPlayer
        from llm_chess_arena.game import Game

        random_player = RandomPlayer(color="black", seed=42)
        game = Game(llm_player_vcr, random_player)

        # Play up to 10 moves
        for _ in range(10):
            if game.board.is_game_over():
                break
            game.make_move()

        # Game should progress without errors
        assert len(game.board.move_stack) >= 1

    @vcr_config.use_cassette("llm_vs_llm_short.yaml")
    def test_llm_vs_llm__with_vcr__then_plays_moves(self):
        """Test LLM vs LLM with recording."""
        connector1 = LLMConnector(model="gpt-4o-mini", temperature=0.0)
        connector2 = LLMConnector(model="gpt-4o-mini", temperature=0.3)
        handler = GameArenaLLMMoveHandler()

        white = LLMPlayer(connector=connector1, handler=handler, color="white")
        black = LLMPlayer(connector=connector2, handler=handler, color="black")

        from llm_chess_arena.game import Game

        game = Game(white, black)

        # Play first 4 moves (2 per side)
        for _ in range(4):
            if game.board.is_game_over():
                break
            game.make_move()

        assert len(game.board.move_stack) >= 2
